refactor(tests): move test files
