+++ e2e/cypress/component/1-store/edges/connection.cy.ts
@@ -1,6 +1,6 @@
 import type { StartHandle } from '@braks/vue-flow'
 import { useVueFlow } from '@braks/vue-flow'
-import { getElements } from '../../../../utils'
+import { getElements } from '../../../utils'
 
 const { nodes, edges } = getElements(2, 2)
