+++ e2e/cypress/component/1-store/edges/removeSelectedEdges.cy.ts
@@ -1,5 +1,5 @@
 import { useVueFlow } from '@braks/vue-flow'
-import { getElements } from '../../../../utils'
+import { getElements } from '../../../utils'
 
 const { nodes, edges } = getElements()
