+++ e2e/cypress/component/1-store/elements/removeSelectedElements.cy.ts
@@ -1,6 +1,6 @@
 import type { GraphEdge, GraphNode } from '@braks/vue-flow'
 import { isNode, useVueFlow } from '@braks/vue-flow'
-import { getElements } from '../../../../utils'
+import { getElements } from '../../../utils'
 
 const { nodes, edges } = getElements()
