+++ e2e/cypress/component/1-store/elements/setElements.cy.ts
@@ -1,5 +1,5 @@
 import { isEdge, isNode, useVueFlow } from '@braks/vue-flow'
-import { getElements } from '../../../../utils'
+import { getElements } from '../../../utils'
 
 const { nodes, edges } = getElements()
